Standardised (corrected) printing of `pid`
There exists no perfect format delegate for `pid_t`, but one thing we can be sure of, the return value of the `fork` manual tells us `pid_t` must be a signed type (as -1 is a potential value). Thus, the unsigned format delegates are probably the choices least portable/most deceptive to us debuggers.