Fix `<ctype.h>` arguments
You need to ensure the arguments passed to `<ctype.h>` functions are `unsigned char` values or `EOF`. Casting to `int` makes no difference here; that conversion would've been implicitly performed anyway due to the type of the function... what you need to do is cast to `unsigned char`.

If you need to understand why this is the case, consider that the argument for character type functions might be used as an index for an array... and negative values would cause negative indexes, leading to a situation where attackers might be able to bleed sensitive parts of program context.